Remove Serialization Guard public APIs - Also refactors reflection helpers into separate file
