+++ src/System.Runtime.Serialization.Formatters/tests/SerializationGuardTests.cs
@@ -73,7 +73,7 @@ namespace System.Runtime.Serialization.Formatters.Tests
             ms.Position = 0;
 
             BinaryFormatter reader = new BinaryFormatter();
-            Assert.Throws<DeserializationBlockedException>(() => reader.Deserialize(ms));
+            Assert.Throws(SerializationGuard.DeserializationBlockedExceptionType, () => reader.Deserialize(ms));
         }
     }
