+++ src/System.Runtime.Serialization.Formatters/tests/System.Runtime.Serialization.Formatters.Tests.csproj
@@ -23,6 +23,9 @@
     <Compile Include="SurrogateSelectorTests.cs" />
     <Compile Include="TargetFrameworkMoniker.cs" />
     <Compile Include="TypeSerializableValue.cs" />
+    <Compile Include="$(CommonPath)\System\Runtime\Serialization\SerializationGuard.cs">
+      <Link>Common\System\Runtime\Serialization\SerializationGuard.cs</Link>
+    </Compile>
     <Compile Include="$(CommonTestPath)\System\NonRuntimeType.cs">
       <Link>Common\System\NonRuntimeType.cs</Link>
     </Compile>
